+++ quantify_scheduler/backends/types/qblox.py
@@ -4,7 +4,7 @@
 
 from __future__ import annotations
 
-from typing import Optional, Dict, Any, Union
+from typing import Optional, Dict, Any, Tuple, Union
 from dataclasses import dataclass
 from dataclasses_json import DataClassJsonMixin
 from quantify_scheduler.backends.qblox import constants
@@ -336,6 +336,8 @@ class SequencerSettings(DataClassJsonMixin):
     """Specifies whether the NCO will be used or not."""
     sync_en: bool
     """Enables party-line synchronization."""
+    connected_outputs: Union[Tuple[int], Tuple[int, int]]
+    """Specifies which physical outputs this sequencer produces waveform data for."""
     modulation_freq: float = None
     """Specifies the frequency of the modulation."""
     mixer_corr_phase_offset_degree: float = 0.0
@@ -349,7 +351,9 @@ class SequencerSettings(DataClassJsonMixin):
 
     @classmethod
     def initialize_from_config_dict(
-        cls, seq_settings: Dict[str, Any]
+        cls,
+        seq_settings: Dict[str, Any],
+        connected_outputs: Union[Tuple[int], Tuple[int, int]],
     ) -> SequencerSettings:
         """
         Instantiates an instance of this class, with initial parameters determined from
@@ -404,6 +408,7 @@ class SequencerSettings(DataClassJsonMixin):
         settings = cls(
             nco_en=nco_en,
             sync_en=True,
+            connected_outputs=connected_outputs,
             modulation_freq=modulation_freq,
             mixer_corr_gain_ratio=mixer_amp_ratio,
             mixer_corr_phase_offset_degree=mixer_phase_error,